Fix memory corruption bug
Gambit's runtime system released the string given to
sqlite3_bind_text() before we had the chance to execute the SQL
statement. strdup() the string to make sure it stays.

TODO: We're replacing a memory corruption with a memory leak. sqlite
does not free the string so we should free it ourselves.